Transport: transport_processor added, part A done.

+++ transport/bus_stop.h
@@ -1,14 +1,14 @@
 #pragma once
 #include <string>
 
-static const double PI = 3.1415926535;
+static const long double PI = 3.1415926535;
 static const unsigned int EARTH_RADIUS = 6'371'000;
 
 struct Coordinates
 {
-    double latitude;
-    double longitude;
-    double DistanceTo(const Coordinates &other) const;
+    long double latitude;
+    long double longitude;
+    long double DistanceTo(const Coordinates &other) const;
 };
 
 struct BusStop
